Moved generic_upgrade_functions.py to root upgrade folder
