+++ upgrade_modules/old_configuration_conversions/alpha_to_beta.py
@@ -2,7 +2,7 @@ from operations_modules import logger
 from operations_modules import file_locations
 from operations_modules.app_generic_functions import get_file_content
 from configuration_modules.config_open_sense_map import CreateOpenSenseMapConfiguration
-from upgrade_modules.old_configuration_conversions.generic_upgrade_functions import reset_primary_config, \
+from upgrade_modules.generic_upgrade_functions import reset_primary_config, \
     reset_installed_sensors, successful_upgrade_message, reset_display_config, reset_trigger_variance_config, \
     reset_mqtt_broker_config, reset_mqtt_publisher_config, reset_mqtt_subscriber_config
